readwritefixed

+++ Askme/app/src/main/java/com/example/dell/askme/QDatabase.java
@@ -9,6 +9,8 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.IgnoreExtraProperties;
 
 import static android.R.attr.name;
+import static com.example.dell.askme.R.id.downvote_count;
+import static com.example.dell.askme.R.id.upvote_count;
 
 @IgnoreExtraProperties
 public class QDatabase {
@@ -19,9 +21,9 @@ public class QDatabase {
     public String option3;
     public String option4;
 
-    public int comm_count;
-    public int upvote_count;
-    public int downvote_count;
+    public long comm_count;
+    public long upvote_count;
+    public long downvote_count;
     String comment;
     private DatabaseReference mDatabase;
 
@@ -31,7 +33,7 @@ public class QDatabase {
     public QDatabase() {
     }
 
-    public QDatabase(String Question, String option1,String option2,String option3,String option4,int upvote_count,int downvote_count,int comm_count,String comment) {
+    public QDatabase(String Question, String option1,String option2,String option3,String option4,long upvote_count,long downvote_count,long comm_count,String comment) {
         this.Question = Question;
         this.option1 = option1;
         this.option2 = option2;
